[server] refactoring
