+++ server/src/auth/mod.rs
@@ -1,2 +1,3 @@
 pub mod filters;
 mod handlers;
+pub mod models;
