+++ server/src/auth/models/mod.rs
@@ -1,3 +1,2 @@
 pub mod session;
-pub mod task;
 pub mod user;
