+++ server/src/auth/models/user.rs
