+++ server/src/tasks/filters.rs
@@ -1,8 +1,8 @@
 use crate::db::PgPool;
-use crate::models::task::ListOptions;
 use crate::routes::utils::{json_body, with_db, with_session_token};
 use crate::tasks::handlers;
 use crate::tasks::handlers::NewTask;
+use crate::tasks::model::ListOptions;
 use warp::Filter;
 
 pub fn routes(
